dom/tests/mochitest/localstorage/frameOrder.html

<html>
<head>
</head>
<script type="text/javascript">
// some tests using this page will trigger an xss violation
netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var prefService = Components.classes["@mozilla.org/preferences-service;1"]
  .getService(Components.interfaces.nsIPrefService);
var cspBranch = prefService.getBranch("security.xssfilter.");
cspBranch.setBoolPref("enabled", false);
</script>

<script type="text/javascript">
function doTest()
{
  var query = location.search.substring(1);
  query = unescape(query);
  var keyNames = eval(query);

  parent.is(localStorage.a, "10", "a = 10");
  parent.is(localStorage.b, "20", "b = 20");
  parent.is(localStorage.c, "30", "c = 30");
  parent.is(localStorage.d, "40", "d = 40");
  parent.is(localStorage.e, "50", "e = 50");
  parent.is(localStorage.length, 5, "length = 5");

  for (var i = 0; i < localStorage.length; ++i)
    parent.is(keyNames[i], localStorage.key(i), "key "+keyNames[i]+" on same index");

  parent.SimpleTest.finish();
  localStorage.clear();
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var prefService = Components.classes["@mozilla.org/preferences-service;1"]
    .getService(Components.interfaces.nsIPrefService);
  var cspBranch = prefService.getBranch("security.xssfilter.");
  cspBranch.setBoolPref("enabled", true);
}
</script>
<body onload="doTest();">
</body>
</html>
